palishopping-kb/requirements.txt
anthropic>=0.86.0
beautifulsoup4>=4.12
google-genai>=1.68.0
httpx>=0.28
lxml>=5.0
Pillow>=11.0
python-dotenv>=1.0
requests>=2.31
rich>=14.0
